#175280704 Get an accommodation and implement restrictions on name, longitude and latitude

+++ src/controllers/accomodation.js
@@ -40,7 +40,13 @@ class Accommodation {
      * step 4: Create accommodation
      */
       const { id, role } = req.user;
-
+      const fac = req.body.facilities;
+      let facilitiesArr = [];
+      if (!Array.isArray(fac)) {
+        facilitiesArr.push(fac);
+      } else {
+        facilitiesArr = facilitiesArr.concat(fac);
+      }
       const user = await users.findOne({ where: { id } });
       let files;
       if (req.files != null) {
@@ -67,22 +73,34 @@ class Accommodation {
               longitude: req.body.longitude,
               latitude: req.body.latitude,
               images: images.map(img => img.url),
-              facilities: req.body.facilities
-            }).then((data) => {
+              facilities: facilitiesArr,
+              hostId: req.user.id,
+              locationId: req.body.locationId
+            }, {
+              include: [
+                {
+                  model: models.location,
+                  as: 'city'
+                }
+              ]
+            }
+            ).then((data) => {
               res.status(201).send({
                 status: 201,
                 data
               });
             }).catch(err => {
-              res.send({ err });
+              console.log(err);
+              res.status(400).send({ error: err });
             });
           } else {
             res.status(403).send({
               status: '403',
-              role: user.role
+              message: 'Unauthorized access'
             });
           }
         }).catch(error => {
+          console.log(error);
           res.send({ error });
         });
     } catch (error) {
@@ -101,10 +119,17 @@ class Accommodation {
        * Deal with images with cloudinary
        * Edit the accomodation
        */
+      const fac = req.body.facilities;
+      const facilitiesArr = [];
+      if (!Array.isArray(fac)) {
+        facilitiesArr.push(fac);
+      } else {
+        facilitiesArr.concat(fac);
+      }
       const { id } = req.user;
       const user = await users.findOne({ where: { id } });
       if (findAccommodation) {
-        if (user.role === 'travel_admin') {
+        if (user.role === 'travel_admin' && findAccommodation.hostId === id) {
           let files;
           if (req.files != null) {
             files = req.files.photo;
@@ -128,7 +153,7 @@ class Accommodation {
                 longitude: req.body.longitude,
                 latitude: req.body.latitude,
                 images: images.map(img => img.url),
-                facilities: req.body.facilities
+                facilities: facilitiesArr
               },
               {
                 where: {
@@ -141,6 +166,11 @@ class Accommodation {
                 res.send({ err });
               });
             });
+        } else {
+          res.status(403).send({
+            status: '403',
+            message: 'Unauthorized access'
+          });
         }
       } else {
         res.status(404).send({
@@ -150,6 +180,31 @@ class Accommodation {
       }
     } catch (error) {
       res.status(500).send({ error });
+    }
+  }
+
+  async getAccommodation(req, res) {
+    const findAccommodation = await accomodation.findOne({ where: { id: req.params.acc_id } });
+    try {
+      if (findAccommodation) {
+        return models.accomodation.findByPk(req.params.acc_id, {
+
+          include: [
+            {
+              model: models.room,
+              as: 'rooms'
+            },
+          ]
+        }).then((info) => {
+          res.status(200).send(info);
+        });
+      }
+      res.status(404).send({
+        status: 404,
+        message: 'Accommodation not found'
+      });
+    } catch (error) {
+      res.status(500).send(error);
     }
   }
 
@@ -165,7 +220,7 @@ class Accommodation {
       const { id } = req.user;
       const user = await users.findOne({ where: { id } });
       if (findAccommodation) {
-        if (user.role === 'travel_admin') {
+        if (user.role === 'travel_admin' && findAccommodation.hostId === id) {
           room.destroy({ where: { accomodationId: req.params.acc_id } });
           return accomodation.destroy({ where: { id: req.params.acc_id } }).then(data => {
             if (data) {
@@ -184,6 +239,10 @@ class Accommodation {
             res.status(409).send(err);
           });
         }
+        res.status(403).send({
+          status: '403',
+          message: 'Unauthorized access'
+        });
       } else {
         res.status(404).send({
           status: 404,
