+++ src/controllers/room.js
@@ -42,7 +42,7 @@ class Room {
       Promise.all(uploadImages)
         .then(images => {
           if (role === 'travel_admin') {
-            if (findAccommodation) {
+            if (findAccommodation && findAccommodation.hostId === id) {
               room.create({
                 type: req.body.type,
                 price: req.body.price,
@@ -61,10 +61,10 @@ class Room {
               }).then((data) => {
                 res.status(201).send(data);
               }).catch(err => {
-                res.send({ err });
+                res.status(501).send({ err });
               });
             } else {
-              res.send({ message: 'Accomodation was not found' });
+              res.status(404).send({ message: 'Accomodation was not found' });
             }
           } else {
             res.status(403).send({
@@ -91,7 +91,7 @@ class Room {
       const { id, role } = req.user;
       const user = await users.findOne({ where: { id } });
       if (findAccommodation && findRoom) {
-        if (role === 'travel_admin') {
+        if (role === 'travel_admin' && findAccommodation.hostId === id) {
           return room.destroy({ where: { id: req.params.room_id } }).then(data => {
             if (data) {
               res.status(200).send({
